feat(aihub): Merge ExpertGroundedAgent into RAGAgent (#681)
Consolidate expert escalation workflow into RAGAgent, providing expert
consultation when context is insufficient after retrieval attempts.

- Add ExpertEscalationConfig for mandatory expert agent configuration
- Add ExpertRejectEvent guard event for user decline handling
- Add UserRequestsExpertEvent control event for consent flow
- Implement 6 new workflow steps in RAGAgent:
  - Ask for consent (with expert_escalation_enabled precondition)
  - Process consent response
  - Forward to ExpertAskingAgent via AgentInTheLoop
  - Handle successful/unsuccessful expert responses
  - Handle expert agent exceptions
- Add i18n translations for ExpertRejectEvent (en, de, fr, it)
- Remove ExpertGroundedAgent (merged functionality)
