--- aihub_agent/aihub_agent/agents/ExpertGroundedAgent/events/UserRequestsExpertEvent.py
@@ -1,9 +1,0 @@
-from aihub_lib.nats.events import ControlEvent
-
-
-class UserRequestsExpertEvent(ControlEvent):
-    """
-    An event signaling that a user has requested an expert's assistance.
-    """
-
-    pass
