+++ aihub_agent/aihub_agent/agents/RagAgent/RAGAgent.py
@@ -12,10 +12,17 @@ from aihub_lib.generative_ai.utils.retrieve_parent_summary_nodes import retrieve
 from aihub_lib.generative_ai.utils.retrieve_prev_next_nodes import retrieve_prev_next_nodes
 from aihub_lib.i18n.LocaleHandler import LocaleHandler
 from aihub_lib.i18n.LocaleString import LocaleString
-from aihub_lib.nats.events import LimitChatHistoryEvent, StandaloneQuestionCondenserEvent
+from aihub_lib.nats.events import (
+    AgentInTheLoop,
+    HumanInTheLoop,
+    LimitChatHistoryEvent,
+    StandaloneQuestionCondenserEvent,
+    StopEvent,
+)
 from aihub_lib.nats.events.guard import (
     ContextInsufficientRejectEvent,
     ContextSufficientAcceptEvent,
+    ExpertRejectEvent,
     FewShotAcceptEvent,
     FewShotRejectEvent,
 )
@@ -27,11 +34,15 @@ from llama_index.core import PromptTemplate
 from llama_index.core.base.llms.types import ChatMessage, MessageRole
 
 from aihub_agent.agents.Agent import Agent
+from aihub_agent.agents.ExpertAskingAgent.events.AnswerStopEvent import AnswerStopEvent
+from aihub_agent.agents.ExpertAskingAgent.events.AskExpertStartEvent import AskExpertStartEvent
+from aihub_agent.agents.ExpertAskingAgent.events.NoAnswerStopEvent import NoAnswerStopEvent
 from aihub_agent.agents.RagAgent.configs.RAGAgentConfig import RAGAgentConfig
 from aihub_agent.agents.RagAgent.configs.RetrieveStepConfig import RetrieveStepConfig
 from aihub_agent.agents.RagAgent.events.ContextInsufficientWithQueryEvent import ContextInsufficientWithQueryEvent
 from aihub_agent.agents.RagAgent.events.InOrderNodeCombinerEvent import InOrderNodeCombinerEvent
 from aihub_agent.agents.RagAgent.events.LimitChatHistoryWithContextEvent import LimitChatHistoryWithContextEvent
+from aihub_agent.agents.RagAgent.events.UserRequestsExpertEvent import UserRequestsExpertEvent
 from aihub_agent.context.run.RunContext import RunContext
 from aihub_agent.workflow.decorators.precondition import precondition
 from aihub_agent.workflow.decorators.step import step
@@ -51,6 +62,24 @@ async def reranking_complete_or_disabled(event: RetrieverEvent | RerankerEvent, 
         return isinstance(event, RetrieverEvent)
     # If reranking is enabled, we must wait for RerankerEvent
     return isinstance(event, RerankerEvent)
+
+
+@precondition()
+async def expert_escalation_enabled(event: ContextInsufficientRejectEvent, config: RAGAgentConfig) -> bool:
+    """Precondition to check if expert escalation is configured."""
+    return config.expert_escalation is not None
+
+
+@precondition()
+async def is_answer_response(event: AgentInTheLoop.response) -> bool:
+    """Ensures agent in the loop response is a successful answer."""
+    return isinstance(event.stop_event, AnswerStopEvent)
+
+
+@precondition()
+async def is_no_answer_response(event: AgentInTheLoop.response) -> bool:
+    """Ensures agent in the loop response is an unsuccessful answer."""
+    return isinstance(event.stop_event, NoAnswerStopEvent)
 
 
 class RAGAgent(Agent):
@@ -326,12 +355,141 @@ class RAGAgent(Agent):
         return LimitChatHistoryWithContextEvent(limited_history_with_context=limited_chat_history)
 
     @step(
+        name=LocaleString(en="Ask for Consent"),
+        description=LocaleString(en="Ask user for consent to contact expert with their question."),
+        icon="akar-icons:chat-approve",
+        precondition=expert_escalation_enabled,
+    )
+    async def insufficient_context_step(
+        self,
+        _: ContextInsufficientRejectEvent,
+        displayer: EventDisplayer,
+        t: LocaleHandler,
+    ) -> HumanInTheLoop.request:
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.context_not_sufficient"))
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.asking_for_consent"))
+        return HumanInTheLoop.invoke(question=t("agent.expert_grounded_agent.messages.consent_question"))
+
+    @step(
+        name=LocaleString(en="Consent Answer"),
+        description=LocaleString(en="User answered the question for consent."),
+        icon="carbon:question-answering",
+    )
+    async def user_expert_inquiry_response(
+        self,
+        event: HumanInTheLoop.response,
+        displayer: EventDisplayer,
+        t: LocaleHandler,
+    ) -> UserRequestsExpertEvent | ExpertRejectEvent:
+        if "yes" in event.response.lower() or "ja" in event.response.lower():
+            await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.user_consented"))
+            return UserRequestsExpertEvent()
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.user_declined"))
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.waiting_for_instructions"))
+        await displayer.display_chunk(
+            t("agent.expert_grounded_agent.messages.user_declined_confirmation"), model_name="gpt-4o"
+        )
+        return ExpertRejectEvent(reason="User declined expert escalation")
+
+    @step(
+        name=LocaleString(en="Invoke ExpertAskingAgent"),
+        description=LocaleString(en="Forwarding request to ExpertAskingAgent that will prompt experts."),
+        icon="hugeicons:robot-02",
+    )
+    async def forward_to_expert_asking_agent_step(
+        self,
+        user_message_event: UserMessageEvent,
+        _: UserRequestsExpertEvent,
+        displayer: EventDisplayer,
+        agent_config: RAGAgentConfig,
+        t: LocaleHandler,
+    ) -> AgentInTheLoop.request:
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.forwarding_to_expert"))
+        await displayer.display_chunk(
+            t("agent.expert_grounded_agent.messages.expert_forwarding_confirmation"), model_name="expert"
+        )
+        await displayer.display_chunk("\n", model_name="expert")
+        await displayer.display_chunk(
+            t("agent.expert_grounded_agent.messages.expert_answer_coming_soon"), model_name="expert"
+        )
+        return AgentInTheLoop.invoke(
+            agent_class=agent_config.expert_escalation.expert_asking_agent_class,
+            agent_id=agent_config.expert_escalation.expert_asking_agent_id,
+            start_event=AskExpertStartEvent(
+                question_to_expert=user_message_event.messages[-1].content,
+                locale=user_message_event.locale,
+                user=user_message_event.user,
+            ),
+        )
+
+    @step(
+        precondition=is_answer_response,
+        name=LocaleString(en="Expert Answer Positive"),
+        description=LocaleString(en="ExpertAskingAgent was able to extract information from expert."),
+        icon="ix:user-success-filled",
+    )
+    async def expert_answered_step(
+        self,
+        displayer: EventDisplayer,
+        event: AgentInTheLoop.response,
+        t: LocaleHandler,
+    ) -> StopEvent:
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.expert_answered"))
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.forwarding_expert_info"))
+        await displayer.display_chunk(event.stop_event.expert_answer, model_name="expert")
+        return StopEvent()
+
+    @step(
+        precondition=is_no_answer_response,
+        name=LocaleString(en="Expert Answer Negative"),
+        description=LocaleString(en="ExpertAskingAgent was NOT able to extract information from expert."),
+        icon="ix:user-fail-filled",
+    )
+    async def expert_not_answered_step(
+        self,
+        displayer: EventDisplayer,
+        _: AgentInTheLoop.response,
+        t: LocaleHandler,
+    ) -> StopEvent:
+        await displayer.display_thought(t("agent.expert_grounded_agent.thoughts.expert_unable_to_answer"))
+        await displayer.display_chunk(
+            t("agent.expert_grounded_agent.messages.expert_unable_to_answer"), model_name="expert"
+        )
+        return StopEvent()
+
+    @step(
+        name=LocaleString(en="Expert Answer Error"),
+        description=LocaleString(en="ExpertAskingAgent encountered an error."),
+        icon="ix:error",
+    )
+    async def expert_exception_step(
+        self,
+        displayer: EventDisplayer,
+        exception_event: AgentInTheLoop.exception,
+        t: LocaleHandler,
+    ) -> StopEvent:
+        await displayer.display_thought(
+            t(
+                "agent.expert_grounded_agent.thoughts.expert_error",
+                error_code=exception_event.exception_event.http_status_code,
+                error_message=exception_event.exception_event.message,
+            )
+        )
+        await displayer.display_chunk(
+            t("agent.expert_grounded_agent.messages.expert_error_occurred"), model_name="expert"
+        )
+        return StopEvent()
+
+    @step(
         name=LocaleString(en="Respond with LLM"),
         description=LocaleString(en="Generates a response using the configured LLM."),
     )
     async def respond_with_llm_step(
         self,
-        event: LimitChatHistoryWithContextEvent | FewShotRejectEvent | ContextInsufficientRejectEvent,
+        event: LimitChatHistoryWithContextEvent
+        | FewShotRejectEvent
+        | ContextInsufficientRejectEvent
+        | ExpertRejectEvent,
         limited_history_without_context: LimitChatHistoryEvent,
         agent_config: RAGAgentConfig,
         displayer: EventDisplayer,
@@ -340,7 +498,7 @@ class RAGAgent(Agent):
         """
         Generates a response using the configured LLM.
         """
-        if isinstance(event, FewShotRejectEvent) or isinstance(event, ContextInsufficientRejectEvent):
+        if isinstance(event, FewShotRejectEvent | ContextInsufficientRejectEvent | ExpertRejectEvent):
             messages = limited_history_without_context.limited_history + [
                 ChatMessage(
                     role=MessageRole.SYSTEM,
