+++ aihub_lib/aihub_lib/i18n/translations/lib/events.en.yml
@@ -90,6 +90,10 @@ context_insufficient_reject_event:
   name: "Insufficient Context Available"
   description: "More context is needed to properly answer the user's query."
 
+expert_reject_event:
+  name: "Expert Escalation Declined"
+  description: "The user declined to escalate their question to a human expert."
+
 sensitive_info_accept_event:
   name: "Response Safe to Share"
   description: "The response does not contain sensitive information and can be safely shared."
